Merge my Miscellaneous lexicon in with main one.
'NIN' gets spoken as "Nine Inch Nails" (at least by Amazon Polly). This grapheme is more for people who use TTT for chat. It may get separated out as an optional lexicon later if we run into the limit at some point.

Regen gets spoken as "reh-gen" so this should make it speak "reejehn". I use this in fights to remind me in solo fights to pop another regen on myself as a WHM. this lexeme may get separated if we run into the Amazon Polly limit.