Added eld work around
Due to eld interrupting words containing the word eld, the entry had been edited and expanded to stop interrupting text with eld, but still allow the Word eld to be pronounced properly with various punctuation. Will need to be a work around for any short word that is contained within other words.