Skip reading voice-acted NPC dialogue
By hooking into FFXIV's sound management, we can detect when
playback starts for voice-acted lines and skip reading out the
corresponding text. Voice lines are detected by applying some heuristics
to their file names; these heuristics may not be perfect. The file name
of a sound is only available when it's loaded, and not later when it's
played back, so some care must be taken to correctly keep track of voice
line sound resources and their corresponding memory addresses as they're
loaded and unloaded by the game.

Resolves #7.